Physics equations Header FIle, Class File, and driver file
This is just a basic header file and cpp file along with a simple driver that i used to test each example in the book. After testing each example i would copy it to the graveyard below in the commented out zone, where it is neatly organized for each problem in the books examples.